LayoutTests/inspector/audit/manager-start-setup-expected.txt

Tests for the AuditManager.prototype.start functionality.


== Running test suite: AuditManager.prototype.start
-- Running test case: AuditManager.prototype.start.SyncSetup
PASS: The setup function should have set test to 42.

-- Running test case: AuditManager.prototype.start.AsyncSetup
PASS: The setup function should have set test to 42.

-- Running test case: AuditManager.prototype.start.SubLevelSetup
PASS: The setup function should have set test to undefined.

-- Running test case: AuditManager.prototype.start.OverriddenSetup
PASS: The setup function should have set test to B.

-- Running test case: AuditManager.prototype.start.MultipleTopLevel
PASS: The setup function should have set test to A.
PASS: The setup function should have set test to B.
